felafax_cli-0.1.17/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
boto3==1.35.56
botocore==1.35.56
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
fastapi==0.115.4
frozenlist==1.5.0
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-compute==1.20.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-cloud-tpu==1.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
keyring==25.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
multidict==6.1.0
nh3==0.2.18
packaging==24.2
pathlib==1.0.1
pkginfo==1.10.0
posthog==3.7.3
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.3
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
twine==5.1.1
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.45.0
yarl==1.18.0
zipp==3.21.0
